requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
boto3==1.35.64
botocore==1.35.64
certifi==2024.8.30
chardet==5.2.0
click==8.1.7
fastapi==0.115.4
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
jmespath==1.0.1
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
reportlab==4.2.5
s3transfer==0.10.3
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
wheel==0.43.0
